Removed deprecated rendering system. Fixed bug in camera projection.
- Updated examples.
