+++ src/pyGandalf/examples/OpenGL/example01.py
@@ -3,13 +3,14 @@ from pyGandalf.core.opengl_window import OpenGLWindow
 from pyGandalf.systems.system import System
 from pyGandalf.systems.link_system import LinkSystem
 from pyGandalf.systems.transform_system import TransformSystem
-from pyGandalf.systems.opengl_rendering_system import OpenGLRenderingSystem
+from pyGandalf.systems.camera_system import CameraSystem
+from pyGandalf.systems.opengl_static_mesh_rendering_system import OpenGLStaticMeshRenderingSystem
 from pyGandalf.renderer.opengl_renderer import OpenGLRenderer
 from pyGandalf.scene.entity import Entity
 from pyGandalf.scene.scene import Scene
 from pyGandalf.scene.scene_manager import SceneManager
 
-from pyGandalf.scene.components import InfoComponent, TransformComponent, LinkComponent, OpenGLRenderComponent, MaterialComponent
+from pyGandalf.scene.components import InfoComponent, TransformComponent, LinkComponent, StaticMeshComponent, MaterialComponent, CameraComponent
 
 from pyGandalf.utilities.opengl_material_lib import OpenGLMaterialLib, MaterialData
 from pyGandalf.utilities.opengl_texture_lib import OpenGLTextureLib
@@ -21,42 +22,58 @@ from pyGandalf.utilities.definitions import SHADERS_PATH, TEXTURES_PATH
 
 import numpy as np
 import glm
+import glfw
 
 """
 Showcase of an ecss cube consisting of an empty parent entity and six other entities for each face of the cube.
 A custom component is added to the root entity to rotate around the whole cube.
 """
 
-class RotateAroundComponent:
-    def __init__(self, axis, speed) -> None:
+class DemoComponent:
+    def __init__(self, axis, speed, rotate_around, main_camera) -> None:
         self.axis = axis
         self.speed = speed
+        self.rotate_around = rotate_around
+        self.main_camera = main_camera
 
-class RotateAroundSystem(System):
+class DemoSystem(System):
     """
-    The system responsible for rotation around.
+    The system responsible showcasing new features.
     """
 
     def on_create(self, entity: Entity, components):
         """
         Gets called once in the first frame for every entity that the system operates on.
         """
-        rotate_around, transform = components
+        demo, transform, info = components
 
     def on_update(self, ts, entity: Entity, components):
         """
         Gets called every frame for every entity that the system operates on.
         """
-        rotate_around, transform = components
+        demo, transform, info = components
 
-        if rotate_around.axis[0] == 1:
-            transform.rotation[0] += rotate_around.speed * ts
+        if demo.rotate_around == True:
+            if demo.axis[0] == 1:
+                transform.rotation[0] += demo.speed * ts
 
-        if rotate_around.axis[1] == 1:
-            transform.rotation[1] += rotate_around.speed * ts
+            if demo.axis[1] == 1:
+                transform.rotation[1] += demo.speed * ts
 
-        if rotate_around.axis[2] == 1:
-            transform.rotation[2] += rotate_around.speed * ts
+            if demo.axis[2] == 1:
+                transform.rotation[2] += demo.speed * ts
+        else:
+            demo.main_camera = False
+
+            if glfw.get_key(Application().get_window().get_handle(), glfw.KEY_C) == glfw.PRESS:
+                if info.tag == 'camera':
+                    demo.main_camera = True
+            elif glfw.get_key(Application().get_window().get_handle(), glfw.KEY_V) == glfw.PRESS:
+                if info.tag == 'camera_alt':
+                    demo.main_camera = True
+
+            SceneManager().get_active_scene().get_component(entity, CameraComponent).primary = demo.main_camera
+        
 
 # Example Usage
 def main():
@@ -65,6 +82,9 @@ def main():
     scene = Scene()
 
     # Create Enroll entities to registry
+    root = scene.enroll_entity()
+    camera = scene.enroll_entity()
+    camera_alt = scene.enroll_entity()
     cube = scene.enroll_entity()
     cube_face_front = scene.enroll_entity()
     cube_face_back = scene.enroll_entity()
@@ -112,59 +132,78 @@ def main():
         [0.0, 1.0]  #0
     ], dtype=np.float32)
 
+
+    scene.add_component(root, TransformComponent(glm.vec3(0, 0, 0), glm.vec3(0, 0, 0), glm.vec3(1, 1, 1)))
+    scene.add_component(root, LinkComponent(None))
+
     # Register components to cube
     scene.add_component(cube, InfoComponent("cube"))
     scene.add_component(cube, TransformComponent(glm.vec3(0, 0, 0), glm.vec3(45, 0, 0), glm.vec3(1, 1, 1)))
-    scene.add_component(cube, LinkComponent(None))
-    scene.add_component(cube, RotateAroundComponent((1, 1, 0), 25))
+    scene.add_component(cube, LinkComponent(root))
+    scene.add_component(cube, DemoComponent((1, 1, 0), 25, True, False))
 
     # Register components to cube_face_front
     scene.add_component(cube_face_front, InfoComponent("cube_face_front"))
     scene.add_component(cube_face_front, TransformComponent(glm.vec3(0, 0, 0.5), glm.vec3(0, 0, 0), glm.vec3(1, 1, 1)))
     scene.add_component(cube_face_front, LinkComponent(cube))
-    scene.add_component(cube_face_front, OpenGLRenderComponent([vertices, texture_coords], None))
+    scene.add_component(cube_face_front, StaticMeshComponent('f', [vertices, texture_coords], None))
     scene.add_component(cube_face_front, MaterialComponent('M_Red_Textured'))
 
     # Register components to cube_face_back
     scene.add_component(cube_face_back, InfoComponent("cube_face_back"))
     scene.add_component(cube_face_back, TransformComponent(glm.vec3(0, 0, -0.5), glm.vec3(0, 0, 0), glm.vec3(1, 1, 1)))
     scene.add_component(cube_face_back, LinkComponent(cube))
-    scene.add_component(cube_face_back, OpenGLRenderComponent([vertices, texture_coords], None))
+    scene.add_component(cube_face_back, StaticMeshComponent('b', [vertices, texture_coords], None))
     scene.add_component(cube_face_back, MaterialComponent('M_Red_Textured'))
 
     # Register components to cube_face_right
     scene.add_component(cube_face_right, InfoComponent("cube_face_right"))
     scene.add_component(cube_face_right, TransformComponent(glm.vec3(0.0, 0, 0.5), glm.vec3(0, 90, 0), glm.vec3(1, 1, 1)))
     scene.add_component(cube_face_right, LinkComponent(cube))
-    scene.add_component(cube_face_right, OpenGLRenderComponent([vertices, texture_coords], None))
+    scene.add_component(cube_face_right, StaticMeshComponent('r', [vertices, texture_coords], None))
     scene.add_component(cube_face_right, MaterialComponent('M_Blue_Textured'))
 
     # Register components to cube_face_left
     scene.add_component(cube_face_left, InfoComponent("cube_face_left"))
     scene.add_component(cube_face_left, TransformComponent(glm.vec3(0.0, 0, -0.5), glm.vec3(0, 90, 0), glm.vec3(1, 1, 1)))
     scene.add_component(cube_face_left, LinkComponent(cube))
-    scene.add_component(cube_face_left, OpenGLRenderComponent([vertices, texture_coords], None))
+    scene.add_component(cube_face_left, StaticMeshComponent('l', [vertices, texture_coords], None))
     scene.add_component(cube_face_left, MaterialComponent('M_Blue_Textured'))
 
     # Register components to cube_face_top
     scene.add_component(cube_face_top, InfoComponent("cube_face_top"))
     scene.add_component(cube_face_top, TransformComponent(glm.vec3(0, 0, 0.5), glm.vec3(90, 0, 0), glm.vec3(1, 1, 1)))
     scene.add_component(cube_face_top, LinkComponent(cube))
-    scene.add_component(cube_face_top, OpenGLRenderComponent([vertices], None))
+    scene.add_component(cube_face_top, StaticMeshComponent('t', [vertices], None))
     scene.add_component(cube_face_top, MaterialComponent('M_Yellow_Simple'))
 
     # Register components to cube_face_bottom
-    scene.add_component(cube_face_bottom, InfoComponent("cube_face_top"))
+    scene.add_component(cube_face_bottom, InfoComponent("cube_face_bottom"))
     scene.add_component(cube_face_bottom, TransformComponent(glm.vec3(0, 0, -0.5), glm.vec3(90, 0, 0), glm.vec3(1, 1, 1)))
     scene.add_component(cube_face_bottom, LinkComponent(cube))
-    scene.add_component(cube_face_bottom, OpenGLRenderComponent([vertices], None))
+    scene.add_component(cube_face_bottom, StaticMeshComponent('b', [vertices], None))
     scene.add_component(cube_face_bottom, MaterialComponent('M_Yellow_Simple'))
+
+    # Register components to camera
+    scene.add_component(camera, InfoComponent("camera"))
+    scene.add_component(camera, TransformComponent(glm.vec3(0, 0, 5), glm.vec3(0, 0, 0), glm.vec3(1, 1, 1)))
+    scene.add_component(camera, LinkComponent(root))
+    scene.add_component(camera, CameraComponent(45, 1.778, 0.1, 1000, 1.2, CameraComponent.Type.PERSPECTIVE))
+    scene.add_component(camera, DemoComponent((1, 1, 0), 25, False, True))
+
+    # Register components to camera_alt
+    scene.add_component(camera_alt, InfoComponent("camera_alt"))
+    scene.add_component(camera_alt, TransformComponent(glm.vec3(0, 0, 5), glm.vec3(0, 0, 0), glm.vec3(1, 1, 1)))
+    scene.add_component(camera_alt, LinkComponent(root))
+    scene.add_component(camera_alt, CameraComponent(45, 1.778, 0.1, 1000, 1.2, CameraComponent.Type.ORTHOGRAPHIC, False))
+    scene.add_component(camera_alt, DemoComponent((1, 1, 0), 25, False, True))
 
     # Create Register systems
     scene.register_system(TransformSystem([TransformComponent]))
     scene.register_system(LinkSystem([LinkComponent, TransformComponent]))
-    scene.register_system(OpenGLRenderingSystem([OpenGLRenderComponent, MaterialComponent, TransformComponent]))
-    scene.register_system(RotateAroundSystem([RotateAroundComponent, TransformComponent]))
+    scene.register_system(CameraSystem([CameraComponent, TransformComponent]))
+    scene.register_system(OpenGLStaticMeshRenderingSystem([StaticMeshComponent, MaterialComponent, TransformComponent]))
+    scene.register_system(DemoSystem([DemoComponent, TransformComponent, InfoComponent]))
 
     # Add scene to manager
     SceneManager().add_scene(scene)
