+++ src/pyGandalf/systems/transform_system.py
@@ -28,17 +28,20 @@ class TransformSystem(System):
         """
         transform = components
 
-        previous_local_matrix = transform.local_matrix
+        if not transform.is_static:
+            previous_local_matrix = transform.local_matrix
 
-        T = glm.translate(glm.mat4(1.0), glm.vec3(transform.translation.x, transform.translation.y, transform.translation.z))
-        R = glm.quat(glm.vec3(glm.radians(transform.rotation.x), glm.radians(transform.rotation.y), glm.radians(transform.rotation.z)))
-        S = glm.scale(glm.mat4(1.0), glm.vec3(transform.scale.x, transform.scale.y, transform.scale.z))
-        
-        transform.quaternion = R
-        transform.local_matrix = S * glm.mat4(R) * T
-        transform.world_matrix = transform.local_matrix
+            T = glm.translate(glm.mat4(1.0), glm.vec3(transform.translation.x, transform.translation.y, transform.translation.z))
+            R = glm.quat(glm.vec3(glm.radians(transform.rotation.x), glm.radians(transform.rotation.y), glm.radians(transform.rotation.z)))
+            S = glm.scale(glm.mat4(1.0), glm.vec3(transform.scale.x, transform.scale.y, transform.scale.z))
+            
+            transform.quaternion = R
+            transform.local_matrix = S * glm.mat4(R) * T
+            transform.world_matrix = transform.local_matrix        
 
-        if (np.array_equal(previous_local_matrix, transform.local_matrix)):
-            transform.is_dirty = False
+            if (np.array_equal(previous_local_matrix, transform.local_matrix)):
+                transform.is_dirty = False
+            else:
+                transform.is_dirty = True
         else:
-            transform.is_dirty = True
+            transform.is_dirty = False
